fix(tests): add tests for WAQ quantities conversion to Spatial blocks
